refactor: restructure UC crates for publishing
