+++ delta-kernel-unity-catalog/src/error.rs
