+++ delta-kernel-unity-catalog/tests/data/catalog_managed_0/part1=0/part-00000-13fefaba-8ec2-4762-b17e-aeda657451c5.c000.snappy.parquet
