+++ unity-catalog-client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "uc-client"
+name = "unity-catalog-client"
 edition.workspace = true
 homepage.workspace = true
 keywords.workspace = true
@@ -9,24 +9,24 @@ readme.workspace = true
 rust-version.workspace = true
 version.workspace = true
 
+[lib]
+name = "unity_catalog_client"
+
 # for cargo-release
 [package.metadata.release]
 release = false
 
 [dependencies]
+delta-kernel-unity-catalog = { path = "../delta-kernel-unity-catalog" }
 reqwest = { version = "0.13", features = ["http2", "json"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tokio = { version = "1", features = ["full"] }
-thiserror = "2.0"
+chrono = { version = "0.4", features = ["serde"] }
 tracing = "0.1"
 url = "2.5"
-chrono = { version = "0.4", features = ["serde"] }
-
-[features]
-default = []
-test-utils = []
 
 [dev-dependencies]
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 clap = { version = "4.5", features = ["derive", "env"] }
+delta_kernel = { path = "../kernel", features = ["arrow-56", "default-engine-rustls", "catalog-managed"] }
+object_store = "0.12"
+tokio = { version = "1", features = ["full"] }
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
