+++ unity-catalog-client/src/tests.rs
@@ -1,17 +1,4 @@
-use crate::models::commits::CommitsRequest;
 use crate::models::credentials::Operation;
-
-#[test]
-fn test_commits_request_builder() {
-    let request = CommitsRequest::new("test-id", "test-uri")
-        .with_start_version(1)
-        .with_end_version(10);
-
-    assert_eq!(request.table_id, "test-id");
-    assert_eq!(request.table_uri, "test-uri");
-    assert_eq!(request.start_version, Some(1));
-    assert_eq!(request.end_version, Some(10));
-}
 
 #[test]
 fn test_operation_display() {
